make playwright happy by moving error formatter tests into their own file
